Added A Temp Win / Loss process
The best place to trigger win or loss is after a death of an actor, whether that be a player character or an enemy, which is why I added this temp logic to this script. This is super simple just to get something that works for EP. I also added a new MonoBehaviour that I intend to move to its' own script tomorrow, I just wrote it here to avoid ".meta" problems. I will also 100% be depreciating the MonoBehaviour after EP in lieu of a more polished level system, death system and etc.